+++ src/Nethermind/Nethermind.Network/Rlpx/NettyFrameEncoder.cs
@@ -40,6 +40,13 @@ namespace Nethermind.Network.Rlpx
 
         protected override void Encode(IChannelHandlerContext context, byte[] message, IByteBuffer output)
         {
+//            string expected = "0000e1c18000000000000000000000000000000000000000000000000000000007c604f04ef90243f9023ff901f9a0ff483e972a04a9a62bb4b7d04ae403c615604e4090521ecc5bb7af67f71be09ca01dcc4de8dec75d7aab85b567b6ccd41ad312451b948a7413f0a142fd40d493479400000042020080a056e81f171bcc55a6ff8345e692c0f86e5b48e01b996cadc001622fb5e363b421ee2100092104b901427600fe0100fe0100fe0100be010024830f424080833d090080010a9883010203a02ba5557a4c62a513c7e56d1bf13373e0da6bec016755483e91589fe1c6d212e2880025252403e8f840df800182520852bf011001808080807e200004c08000000000000000000000000000000000000000000000000000000000000000";
+//            string result__ = "0000e1c18000000000000000000000000000000000000000000000000000000001c604f04ef90243f9023ff901f9a0ff483e972a04a9a62bb4b7d04ae403c615604e4090521ecc5bb7af67f71be09ca01dcc4de8dec75d7aab85b567b6ccd41ad312451b948a7413f0a142fd40d493479400000042020080a056e81f171bcc55a6ff8345e692c0f86e5b48e01b996cadc001622fb5e363b421ee2100092104b901427600fe0100fe0100fe0100be010024830f424080833d090080010a9883010203a02ba5557a4c62a513c7e56d1bf13373e0da6bec016755483e91589fe1c6d212e2880025252403e8f840df800182520852bf011001808080807e200004c08000000000000000000000000000000000000000000000000000000000000000";
+//            if (message.ToHexString() != expected)
+//            {
+//                throw new Exception(message.ToHexString());
+//            }
+            
             if (message.Length % 16 != 0)
             {
                 throw new InvalidOperationException($"Frame length should be a multiple of 16");
